+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## [Unreleased]
 - refactoring the internal infoview, and
 - add a first implementation for trace message in internal infoview
+- add a first implementation for goto definition for vscode-adapted infoview
+- (no code changed but two performance related issue: [ref])
 
 ## [0.1.6] - 2024-11-25